+++ CHANGELOG.md
@@ -6,7 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 #### Added
+- **[MODULE]** Update Frida based SSL pinning bypass with iOS10 support _[from @leonjza]_
+
 #### Fixed
+
 #### Removed
 
 
@@ -15,8 +18,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 #### Fixed
 - **[CORE]** Search PID for System Apps
 - **[MODULE]** Keychain extraction of data not encodable in UTF8 _[from @federicodotta]_
-- **[MODULE]** Improved jailbreak detection bypass (`dynamic/detection/script_jailbreak-detection-bypass.py`)
-- **[MODULE]** Improved certificate pinning bypass (`comms/proxy/pinning_bypass_frida`)